Update Events.md
Handlers spilt into ENV, Firefly and not supported. Also added handlers not in the list yet. Don't know (yet) if they are from previous Magic versions ( < XPA, like M9, UP, eDeveloper) or newly added to Firefly? I'm not sure everything is correct. The git editor has a glitch when you insert something into an existing row. Values are then doubled and/or overwritten it seems. I also don't have the migration engine code to check if everything is really correct or not. So this is the best I can do for now. Better than nothing!